+++ R/arg-checks.R
@@ -20,21 +20,49 @@ stopifnot_string <- function(x) {
 }
 
 # Checks if argument is an integer
-stopifnot_int <- function(x) {
+stopifnot_int <- function(x, lb = NULL, ub = NULL) {
   if (!is.numeric(x) || !all.equal(x, as.integer(x))) {
     stop(
       paste(match.call()$x, "must be an integer, but is of class(es): "),
       paste(class(x), collapse = ", ")
     )
   }
+
+  if (!is.null(lb) && any(x < lb)) {
+    stop(
+      paste(match.call()$x, "must be greater than or equal to", lb, ", but is:"),
+      paste(x[x < lb], collapse = ", ")
+    )
+  }
+
+  if (!is.null(ub) && any(x > ub)) {
+    stop(
+      paste(match.call()$x, "must be less than or equal to", ub, ", but is:"),
+      paste(x[x > ub], collapse = ", ")
+    )
+  }
 }
 
 # Checks if argument is a double
-stopifnot_double <- function(x) {
+stopifnot_double <- function(x, lb = NULL, ub = NULL) {
   if (!is.numeric(x)) {
     stop(
       paste(match.call()$x, "must be a double, but is of class(es): "),
       paste(class(x), collapse = ", ")
+    )
+  }
+
+  if (!is.null(lb) && any(x < lb)) {
+    stop(
+      paste(match.call()$x, "must be greater than or equal to", lb, ", but is:"),
+      paste(x[x < lb], collapse = ", ")
+    )
+  }
+
+  if (!is.null(ub) && any(x > ub)) {
+    stop(
+      paste(match.call()$x, "must be less than or equal to", ub, ", but is:"),
+      paste(x[x > ub], collapse = ", ")
     )
   }
 }