+++ inst/include/epiworld/models/measlesquarantine.hpp
@@ -24,6 +24,16 @@
  * @brief Template for a Measles model with quarantine
  * 
  * @param TSeq The type of the sequence to be used.
+ * @details
+ * This model can be described as a SEIHR model with isolation and quarantine.
+ * The infectious state is divided into prodromal and rash phases. Furthermore,
+ * the quarantine state includes exposed, susceptible, prodromal, and recovered
+ * states.
+ * 
+ * The quarantine process is triggered any time that an agent with rash is
+ * detected. The agent is then isolated and all agents who are unvaccinated are
+ * quarantined. Isolated agents then may be moved out of the isolation in 
+ * isolation_period days.
  */
 template<typename TSeq = EPI_DEFAULT_TSEQ>
 class ModelMeaslesQuarantine: public Model<TSeq> {
@@ -39,6 +49,7 @@ private:
     static void m_update_prodromal(Agent<TSeq> * p, Model<TSeq> * m);
     static void m_update_rash(Agent<TSeq> * p, Model<TSeq> * m);
     static void m_update_isolated(Agent<TSeq> * p, Model<TSeq> * m);
+    static void m_update_isolated_recovered(Agent<TSeq> * p, Model<TSeq> * m);
     static void m_update_q_exposed(Agent<TSeq> * p, Model<TSeq> * m);
     static void m_update_q_susceptible(Agent<TSeq> * p, Model<TSeq> * m);
     static void m_update_q_prodromal(Agent<TSeq> * p, Model<TSeq> * m);
@@ -60,12 +71,13 @@ public:
     static const epiworld_fast_uint PRODROMAL               = 2u;
     static const epiworld_fast_uint RASH                    = 3u;
     static const epiworld_fast_uint ISOLATED                = 4u;
-    static const epiworld_fast_uint QUARANTINED_EXPOSED     = 5u;
-    static const epiworld_fast_uint QUARANTINED_SUSCEPTIBLE = 6u;
-    static const epiworld_fast_uint QUARANTINED_PRODROMAL   = 7u;
-    static const epiworld_fast_uint QUARANTINED_RECOVERED   = 8u;
-    static const epiworld_fast_uint HOSPITALIZED            = 9u;
-    static const epiworld_fast_uint RECOVERED               = 10u;
+    static const epiworld_fast_uint ISOLATED_RECOVERED      = 5u;
+    static const epiworld_fast_uint QUARANTINED_EXPOSED     = 6u;
+    static const epiworld_fast_uint QUARANTINED_SUSCEPTIBLE = 7u;
+    static const epiworld_fast_uint QUARANTINED_PRODROMAL   = 8u;
+    static const epiworld_fast_uint QUARANTINED_RECOVERED   = 9u;
+    static const epiworld_fast_uint HOSPITALIZED            = 10u;
+    static const epiworld_fast_uint RECOVERED               = 11u;
     
     // Default constructor
     ModelMeaslesQuarantine() {};
@@ -82,10 +94,11 @@ public:
      * @param rash_period The rash period of the virus.
      * @param days_undetected The number of days the virus goes undetected.
      * @param hospitalization_rate The rate of hospitalization.
-     * @param hospitalization_duration The duration of hospitalization.
+     * @param hospitalization_period The duration of hospitalization.
      * @param prop_vaccinated The proportion of vaccinated agents.
-     * @param quarantine_days The number of days for quarantine.
+     * @param quarantine_period The number of days for quarantine.
      * @param quarantine_willingness The willingness to be quarantined.
+     * @param isolation_period The number of days for isolation.
      */
     ///@{ 
     ModelMeaslesQuarantine(
@@ -102,11 +115,12 @@ public:
         epiworld_double rash_period,
         epiworld_double days_undetected,
         epiworld_double hospitalization_rate,
-        epiworld_double hospitalization_duration,
+        epiworld_double hospitalization_period,
         // Policy parameters
         epiworld_double prop_vaccinated,
-        epiworld_fast_uint quarantine_days,
-        epiworld_double quarantine_willingness
+        epiworld_fast_uint quarantine_period,
+        epiworld_double quarantine_willingness,
+        epiworld_fast_uint isolation_period
     );
 
     ModelMeaslesQuarantine(
@@ -122,11 +136,12 @@ public:
         epiworld_double rash_period,
         epiworld_double days_undetected,
         epiworld_double hospitalization_rate,
-        epiworld_double hospitalization_duration,
+        epiworld_double hospitalization_period,
         // Policy parameters
         epiworld_double prop_vaccinated,
-        epiworld_fast_uint quarantine_days,
-        epiworld_double quarantine_willingness
+        epiworld_fast_uint quarantine_period,
+        epiworld_double quarantine_willingness,
+        epiworld_fast_uint isolation_period
     );
     ///@}
 
@@ -135,6 +150,7 @@ public:
     bool system_quarantine_triggered = false;
 
     std::vector< int > day_flagged; ///< Either detected or started quarantine
+    std::vector< int > day_rash_onset; ///< Day of rash onset
 
     /**
      * @brief Quarantine agents that are in the system.
@@ -170,6 +186,13 @@ inline void ModelMeaslesQuarantine<TSeq>::quarantine_agents() {
     if (!system_quarantine_triggered)
         return;
 
+    // Quarantine and isolation can be shut off if negative
+    if (
+        (this->par("Quarantine period") < 0) &&
+        (this->par("Isolation period") < 0)
+    )
+        return;
+
     // Capturing the days that matter and the probability of success
     epiworld_double willingness = this->par("Quarantine willingness");
 
@@ -184,7 +207,7 @@ inline void ModelMeaslesQuarantine<TSeq>::quarantine_agents() {
 
         // Rash is evident, so they go straight to isolation, in all
         // cases.
-        if (agent_state == RASH)
+        if ((agent_state == RASH) && (this->par("Isolation period") >= 0))
         {
             this->get_agent(i).change_state(this, ISOLATED);
             this->day_flagged[i] = this->today();
@@ -197,7 +220,10 @@ inline void ModelMeaslesQuarantine<TSeq>::quarantine_agents() {
 
         // Quarantine will depend on the willingness of the agent
         // to be quarantined. If negative, then quarantine never happens.
-        if (this->runif() < willingness)
+        if (
+            (this->par("Quarantine period") >= 0) &&
+            (this->runif() < willingness)
+        )
         {
 
             if (agent_state == SUSECPTIBLE)
@@ -246,6 +272,12 @@ inline void ModelMeaslesQuarantine<TSeq>::reset() {
         day_flagged.end(),
         0);
 
+    this->day_rash_onset.resize(this->size(), 0);
+    std::fill(
+        day_rash_onset.begin(),
+        day_rash_onset.end(),
+        0);
+
     this->m_update_model(dynamic_cast<Model<TSeq>*>(this));
     return;
     
@@ -258,7 +290,7 @@ inline void ModelMeaslesQuarantine<TSeq>::update_available() {
     for (auto & agent: this->get_agents())
     {
         const auto & s = agent.get_state();
-        if (s <= RASH)
+        if (s < RASH)
             this->available.push_back(&agent);
     }
 
@@ -327,11 +359,8 @@ LOCAL_UPDATE_FUN(m_update_susceptible) {
         if (neighbor.get_virus() == nullptr)
             continue;
 
-        // Only infectious individuals can transmit
-        if (
-            (neighbor.get_state() != model->PRODROMAL) &&
-            (neighbor.get_state() != model->RASH)
-        )
+        // Only prodomal individuals can transmit
+        if (neighbor.get_state() != model->PRODROMAL)
             continue;
 
         auto & v = neighbor.get_virus();
@@ -380,7 +409,13 @@ LOCAL_UPDATE_FUN(m_update_exposed) {
 LOCAL_UPDATE_FUN(m_update_prodromal) {
     
     if (m->runif() < (1.0/m->par("Prodromal period")))
+    {
+
+        GET_MODEL(model, m);
+        model->day_rash_onset[p->get_id()] = m->today();
         p->change_state(m, ModelMeaslesQuarantine<TSeq>::RASH);
+
+    }
 
     return;
 
@@ -406,7 +441,7 @@ LOCAL_UPDATE_FUN(m_update_rash) {
     // How many days since detected  
     bool detected = false;
     if (
-        (m->par("Days undetected") >= 0) &&
+        (m->par("Isolation period") >= 0) &&
         (m->runif() < 1.0/m->par("Days undetected"))
     )
     {
@@ -417,16 +452,8 @@ LOCAL_UPDATE_FUN(m_update_rash) {
 
     // Probability of Staying in the rash period vs becoming
     // hospitalized
-    m->array_double_tmp[0] = 1.0/m->par("Rash days");
+    m->array_double_tmp[0] = 1.0/m->par("Rash period");
     m->array_double_tmp[1] = m->par("Hospitalization rate");
-
-    #ifdef EPI_DEBUG
-    EPI_DEBUG_PRINTF(
-        "Rash days: %f, Hospitalization rate: %f\n",
-        m->array_double_tmp[0],
-        m->array_double_tmp[1]
-    );
-    #endif
 
     // Sampling from the probabilities
     SAMPLE_FROM_PROBS(2, which);
@@ -439,7 +466,7 @@ LOCAL_UPDATE_FUN(m_update_rash) {
             model->day_flagged[p->get_id()] = m->today();
             p->rm_agent_by_virus(
                 m,
-                ModelMeaslesQuarantine::QUARANTINED_RECOVERED
+                ModelMeaslesQuarantine::ISOLATED_RECOVERED
             );
 
             // NOT TRACKING WHEN THIS AGENT MOVES OUT OF RECOVERED
@@ -467,15 +494,15 @@ LOCAL_UPDATE_FUN(m_update_isolated) {
 
     // Figuring out if the agent can be released from isolation
     // if the quarantine period is over.
-    int days_since = m->today() - model->day_flagged[p->get_id()];
+    int days_since = m->today() - model->day_rash_onset[p->get_id()];
 
-    bool unquarantine =
-        (m->par("Quarantine days") <= days_since) ?
+    bool unisolate =
+        (m->par("Isolation period") <= days_since) ?
         true: false;
 
     // Probability of staying in the rash period vs becoming
     // hospitalized
-    m->array_double_tmp[0] = 1.0/m->par("Rash days");
+    m->array_double_tmp[0] = 1.0/m->par("Rash period");
     m->array_double_tmp[1] = m->par("Hospitalization rate");
 
     // Sampling from the probabilities
@@ -484,9 +511,8 @@ LOCAL_UPDATE_FUN(m_update_isolated) {
     // Recovers
     if (which == 2u)
     {
-        if (unquarantine)
+        if (unisolate)
         {
-            model->day_flagged[p->get_id()] = m->today();
             p->rm_agent_by_virus(
                 m,
                 ModelMeaslesQuarantine::RECOVERED
@@ -494,7 +520,7 @@ LOCAL_UPDATE_FUN(m_update_isolated) {
         }
         else
             p->rm_agent_by_virus(
-                m, ModelMeaslesQuarantine::QUARANTINED_RECOVERED
+                m, ModelMeaslesQuarantine::ISOLATED_RECOVERED
             );
     }
 
@@ -505,10 +531,27 @@ LOCAL_UPDATE_FUN(m_update_isolated) {
     }
     // If neither hospitalized nor recovered, then the agent is
     // still under isolation, unless the quarantine period is over.
-    else if ((which == 0u) && unquarantine)
+    else if ((which == 0u) && unisolate)
     {
         p->change_state(m, ModelMeaslesQuarantine::RASH);
     }
+
+}
+
+LOCAL_UPDATE_FUN(m_update_isolated_recovered) {
+
+    GET_MODEL(model, m);
+
+    // Figuring out if the agent can be released from isolation
+    // if the quarantine period is over.
+    int days_since = m->today() - model->day_rash_onset[p->get_id()];
+
+    bool unisolate =
+        (m->par("Isolation period") <= days_since) ?
+        true: false;
+
+    if (unisolate)
+        p->change_state(m, ModelMeaslesQuarantine::RECOVERED);
 
 }
 
@@ -520,7 +563,7 @@ LOCAL_UPDATE_FUN(m_update_q_exposed) {
         m->today() - model->day_flagged[p->get_id()];
 
     bool unquarantine =
-        (m->par("Quarantine days") <= days_since) ?
+        (m->par("Quarantine period") <= days_since) ?
         true: false;
 
     // Will develop prodromal symptoms?
@@ -557,7 +600,7 @@ LOCAL_UPDATE_FUN(m_update_q_susceptible) {
     int days_since =
         m->today() - model->day_flagged[p->get_id()];
     
-    if (days_since >= m->par("Quarantine days"))
+    if (days_since >= m->par("Quarantine period"))
         p->change_state(m, ModelMeaslesQuarantine::SUSECPTIBLE);
 
 }
@@ -571,7 +614,7 @@ LOCAL_UPDATE_FUN(m_update_q_prodromal) {
     int days_since = m->today() - model->day_flagged[p->get_id()];
 
     bool unquarantine =
-        (m->par("Quarantine days") <= days_since) ?
+        (m->par("Quarantine period") <= days_since) ?
         true: false;
     
     // If they develop rash, then they are isolated and contact
@@ -584,10 +627,10 @@ LOCAL_UPDATE_FUN(m_update_q_prodromal) {
     else
     {
 
-        if (unquarantine)
-            p->change_state(m, ModelMeaslesQuarantine::RASH);
-        else
-            p->change_state(m, ModelMeaslesQuarantine::ISOLATED);        
+        // If develops rash during quarantine, they are moved to 
+        // isolation right away.
+        p->change_state(m, ModelMeaslesQuarantine::ISOLATED);        
+        model->day_rash_onset[p->get_id()] = m->today();
 
     }
 
@@ -598,7 +641,7 @@ LOCAL_UPDATE_FUN(m_update_q_recovered) {
     GET_MODEL(model, m);
     int days_since = m->today() - model->day_flagged[p->get_id()];
     
-    if (days_since >= m->par("Quarantine days"))
+    if (days_since >= m->par("Quarantine period"))
         p->change_state(m, ModelMeaslesQuarantine::RECOVERED);
 
 }
@@ -606,7 +649,7 @@ LOCAL_UPDATE_FUN(m_update_q_recovered) {
 LOCAL_UPDATE_FUN(m_update_hospitalized) {
 
     // The agent is removed from the system
-    if (m->runif() < 1.0/m->par("Hospitalization days"))
+    if (m->runif() < 1.0/m->par("Hospitalization period"))
         p->rm_agent_by_virus(m, ModelMeaslesQuarantine::RECOVERED);
 
     return;
@@ -629,11 +672,12 @@ inline ModelMeaslesQuarantine<TSeq>::ModelMeaslesQuarantine(
     epiworld_double rash_period,
     epiworld_double days_undetected,
     epiworld_double hospitalization_rate,
-    epiworld_double hospitalization_duration,
+    epiworld_double hospitalization_period,
     // Policy parameters
     epiworld_double prop_vaccinated,
-    epiworld_fast_uint quarantine_days,
-    epiworld_double quarantine_willingness
+    epiworld_fast_uint quarantine_period,
+    epiworld_double quarantine_willingness,
+    epiworld_fast_uint isolation_period
 ) {
 
     model.add_state("Susceptible", this->m_update_susceptible);
@@ -641,6 +685,9 @@ inline ModelMeaslesQuarantine<TSeq>::ModelMeaslesQuarantine(
     model.add_state("Prodromal", this->m_update_prodromal);
     model.add_state("Rash", this->m_update_rash);
     model.add_state("Isolated", this->m_update_isolated);
+    model.add_state(
+        "Isolated Recovered", this->m_update_isolated_recovered
+    );
     model.add_state(
         "Quarantined Exposed", this->m_update_q_exposed
     );
@@ -666,14 +713,15 @@ inline ModelMeaslesQuarantine<TSeq>::ModelMeaslesQuarantine(
     model.add_param(transmission_rate, "Transmission rate");
     model.add_param(incubation_period, "Incubation period");
     model.add_param(prodromal_period, "Prodromal period");
-    model.add_param(rash_period, "Rash days");
+    model.add_param(rash_period, "Rash period");
     model.add_param(days_undetected, "Days undetected");
-    model.add_param(quarantine_days, "Quarantine days");
+    model.add_param(quarantine_period, "Quarantine period");
     model.add_param(
         quarantine_willingness, "Quarantine willingness"
     );
+    model.add_param(isolation_period, "Isolation period");
     model.add_param(hospitalization_rate, "Hospitalization rate");
-    model.add_param(hospitalization_duration, "Hospitalization days");
+    model.add_param(hospitalization_period, "Hospitalization period");
     model.add_param(prop_vaccinated, "Vaccination rate");
     model.add_param(vax_efficacy, "Vax efficacy");
     model.add_param(vax_reduction_recovery_rate, "Vax improved recovery");
@@ -682,7 +730,7 @@ inline ModelMeaslesQuarantine<TSeq>::ModelMeaslesQuarantine(
     Virus<> measles("Measles");
     measles.set_state(EXPOSED, RECOVERED);
     measles.set_prob_infecting(&model("Transmission rate"));
-    measles.set_prob_recovery(&model("Rash days"));
+    measles.set_prob_recovery(&model("Rash period"));
     measles.set_incubation(&model("Incubation period"));
     measles.set_distribution(
         distribute_virus_randomly(n_exposed, false)
@@ -725,11 +773,13 @@ inline ModelMeaslesQuarantine<TSeq>::ModelMeaslesQuarantine(
     epiworld_double rash_period,
     epiworld_double days_undetected,
     epiworld_double hospitalization_rate,
-    epiworld_double hospitalization_duration,
+    epiworld_double hospitalization_period,
     // Policy parameters
     epiworld_double prop_vaccinated,
-    epiworld_fast_uint quarantine_days,
-    epiworld_double quarantine_willingness
+    epiworld_fast_uint quarantine_period,
+    epiworld_double quarantine_willingness,
+    epiworld_fast_uint isolation_period
+
 ) {
 
     ModelMeaslesQuarantine(
@@ -745,10 +795,11 @@ inline ModelMeaslesQuarantine<TSeq>::ModelMeaslesQuarantine(
         rash_period,
         days_undetected,
         hospitalization_rate,
-        hospitalization_duration,
+        hospitalization_period,
         prop_vaccinated,
-        quarantine_days,
-        quarantine_willingness
+        quarantine_period,
+        quarantine_willingness,
+        isolation_period
     );
 
     return;